devpro/requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
djangorestframework==3.11.0
importlib-metadata==1.6.0
kombu==4.6.8
psycopg2==2.8.4
pytz==2019.3
radius==0.0.4
redis==3.4.1
redis-server==5.0.7
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
psycopg2-binary==2.8.3
django-admin-rangefilter